Dashboard API: add data transfer as GPX and removed unnecessary IntentChooser.
