+++ src/main/java/de/dennisguse/opentracks/TrackRecordedActivity.java
@@ -39,6 +39,7 @@ import de.dennisguse.opentracks.fragments.ConfirmDeleteDialogFragment;
 import de.dennisguse.opentracks.fragments.StatisticsRecordedFragment;
 import de.dennisguse.opentracks.services.TrackRecordingServiceConnection;
 import de.dennisguse.opentracks.settings.SettingsActivity;
+import de.dennisguse.opentracks.util.IntentDashboardUtils;
 import de.dennisguse.opentracks.util.IntentUtils;
 import de.dennisguse.opentracks.util.PreferencesUtils;
 import de.dennisguse.opentracks.util.TrackIconUtils;
@@ -186,7 +187,7 @@ public class TrackRecordedActivity extends AbstractListActivity implements Choos
                 startActivity(intent);
                 return true;
             case R.id.track_detail_menu_show_on_map:
-                IntentUtils.showTrackOnMapDashboard(this, new long[]{trackId});
+                IntentDashboardUtils.startDashboard(this, trackId);
                 return true;
             case R.id.track_detail_markers:
                 intent = IntentUtils.newIntent(this, MarkerListActivity.class)
