+++ src/main/res/values/strings.xml
@@ -96,7 +96,8 @@ limitations under the License.
     </string>
 
     <string name="open_track_using_dashboard">Open using OpenTracks-Dashboard</string>
-    <string name="open_track_as_file">Open as file</string>
+    <string name="open_track_as_KMZ">Open as KMZ</string>
+    <string name="open_track_as_GPX">Open as GPX</string>
     <string name="open_track_as_trackfileformat">Open as %1$s</string>
 
     <!-- Activity Type -->
